Introduce Application class which handles all high level event processing.
