+++ src/Visual.cpp
@@ -6,6 +6,7 @@
  */
 
 #include "Visual.h"
+#include "GraphicsModel.h"
 
 namespace rose {
 
@@ -14,19 +15,19 @@ namespace rose {
     SemanticGesture SemanticGesture::Scroll{0x2};
     SemanticGesture SemanticGesture::Multi{0x4};
 
-    void rose::Screen::draw(const Position &containerPosition) {
+    void rose::Screen::draw(gm::Context &context, const Position &containerPosition) {
         setScreenRectangle(containerPosition);
         for (auto &content : (*this)) {
             if (auto window = std::dynamic_pointer_cast<Window>(content); window) {
-                window->draw(mScreenRect.position());
+                window->draw(context, mScreenRect.position());
             }
         }
     }
 
-    Rectangle Screen::layout(const Rectangle &screenRect) {
+    Rectangle Screen::layout(gm::Context &context, const Rectangle &screenRect) {
         for (auto &content : (*this)) {
             if (auto window = std::dynamic_pointer_cast<Window>(content); window) {
-                window->layout(screenRect);
+                window->layout(context, screenRect);
             }
         }
         return screenRect;
@@ -50,36 +51,36 @@ namespace rose {
         return std::nullopt;
     }
 
-    void Window::draw(const Position &containerPosition) {
+    void Window::draw(gm::Context &context, const Position &containerPosition) {
         setScreenRectangle(containerPosition);
         for (auto &content : (*this)) {
             if (auto manager = std::dynamic_pointer_cast<Manager>(content); manager) {
-                manager->draw(mScreenRect.position());
+                manager->draw(context, mScreenRect.position());
             }
         }
     }
 
-    Rectangle Window::layout(const Rectangle &screenRect) {
+    Rectangle Window::layout(gm::Context &context, const Rectangle &screenRect) {
         for (auto &content : (*this)) {
             if (auto window = std::dynamic_pointer_cast<Window>(content); window) {
-                window->layout(screenRect);
+                window->layout(context, screenRect);
             }
         }
         return screenRect;
     }
 
-    void Manager::draw(const Position &containerPosition) {
+    void Manager::draw(gm::Context &context, const Position &containerPosition) {
         setScreenRectangle(containerPosition);
         for (auto &content : (*this)) {
             if (auto manager = std::dynamic_pointer_cast<Manager>(content); manager) {
-                manager->draw(mScreenRect.position());
+                manager->draw(context, mScreenRect.position());
             } else if (auto widget = std::dynamic_pointer_cast<Widget>(content); widget) {
-                widget->draw(mScreenRect.position());
+                widget->draw(context, mScreenRect.position());
             }
         }
     }
 
-    Rectangle Manager::layout(const Rectangle &screenRect) {
+    Rectangle Manager::layout(gm::Context &context, const Rectangle &screenRect) {
         Rectangle managerRect{screenRect};
         if (mPreferredPos != Position::Zero) {
             managerRect = mPreferredPos;
@@ -89,7 +90,7 @@ namespace rose {
             managerRect = mPreferredSize;
         }
 
-        auto rect = mLayoutManager->layoutContent(managerRect, begin(), end());
+        auto rect = mLayoutManager->layoutContent(context, managerRect, begin(), end());
         mPos = rect.position();
         mSize = rect.size();
         return rect;
@@ -112,18 +113,19 @@ namespace rose {
     }
 
     Rectangle
-    SimpleLayout::layoutContent(const Rectangle &screenRect, LayoutManager::Itr first, LayoutManager::Itr last) {
+    SimpleLayout::layoutContent(gm::Context &context, const Rectangle &screenRect, LayoutManager::Itr first,
+                                LayoutManager::Itr last) {
         while (first != last) {
             if (auto manager = std::dynamic_pointer_cast<Manager>(*first); manager) {
-                manager->layout(screenRect);
+                manager->layout(context, screenRect);
             } else if (auto widget = std::dynamic_pointer_cast<Widget>(*first); widget) {
-                widget->layout(screenRect);
+                widget->layout(context, screenRect);
             }
         }
         return screenRect;
     }
 
-    void Widget::draw(const Position &containerPosition) {
+    void Widget::draw(gm::Context &context, const Position &containerPosition) {
 
     }
 }
